getSelectedElementSet now handles parameters.
passes tests 1, 2, 3, 5 in Section 2.3
An additional parameter `refpoint` indicates at which location/corner the selector is attached to the instruction. This determines which corners of the selector rectangle pick which dimensions of the parameter object. `pX` is the corner offset in X direction from `refpoint`. `pY` is its Y equivalent. `pXY` is diagonally adjacent from `refpoint`.